requirements.txt
PyGithub==1.54.1
atlassian-python-api==2.5.0
requests==2.25.1

# dev
ipykernel==5.4.2
black==20.8b1
flake8==3.8.4
pytest==6.2.1
